SpeakerRecognition/requirements.txt
numpy==1.14.0
PyAudio==0.2.11
python-speech-features==0.6
redis==3.0.1
scikit-learn==0.19.0
scipy==1.1.0
webrtcvad==2.0.10
